refactored login menu button
